+++ HISTORY.rst
@@ -4,12 +4,13 @@ Changelog
 2.0 (unreleased)
 ****************
 
+- Added session recorder (``doitlive record``).
 - Improved interface.
 - Sessions are played with ``doitlive play <session_file>``.
 - Deprecated ``doitlive-demo``. Run ``doitlive demo`` instead.
 - Deprecated ``doitlive --themes`` and ``doitlive --themes-preview``. Run ``doitlive themes`` and ``doitlive themes --preview`` instead.
-- Added session recorder (``doitlive record``).
 - Fix bug that raised an error when cd'ing into a non-existent directory.
+- Add 'robbyrussell' theme.
 
 
 1.0 (2014-06-18)